safari-600.7-branch/LayoutTests/js/dom/dfg-typed-array-neuter-expected.txt

Tests that DFG respects neutered typed arrays.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS array.length is 100
PASS array[0] is 42
PASS foo(array) is 100
PASS bar(array) is 42
PASS array.length is 0
PASS array[0] is void 0
PASS foo(array) is 0
PASS bar(array) is void 0
PASS successfullyParsed is true

TEST COMPLETE
